requirements.txt
pyarrow==16.1.0
minio
ccxt==4.1.100
apache-airflow-providers-mongo==3.5.0
airflow-provider-great-expectations==0.2.7
pyspark==3.5.5
apache-airflow-providers-apache-spark
psycopg2-binary